Create README.md
Improve functions to make larger pages easier and keep code cleaner
Creating templates functions in javascript to make it easy and clean to build html pages using "kendo js". To make the construction of the pages cleaner, we decided to create a template system so that the business engineer can simply create inputs in a simpler way.